Split PrepAnnotationJson into two methods in order to parallelize
