+++ .dockstore.yml
@@ -143,6 +143,7 @@ workflows:
          - master
          - ah_var_store
          - rc-vs-651-vat-from-vds
+         - gg3_rc-vs-651-vat-from-vds
    - name: GvsValidateVat
      subclass: WDL
      primaryDescriptorPath: /scripts/variantstore/variant_annotations_table/GvsValidateVAT.wdl
